refactor: router 위치 변경
